requirements.txt
streamlit>=1.36.0
pydantic>=2.5.0
pyyaml>=6.0
pytest>=7.4.0
huggingface-hub>=0.20.0
gradio_client>=1.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
